+++ src/main/java/com/binance/api/client/json/Decoders.scala
@@ -1,0 +1,3 @@
+package com.binance.api.client.json
+
+object Decoders extends AccountDecoders with EventDecoders with GeneralDecoders with MarketDecoders
