+++ src/main/java/com/binance/api/client/json/DomainDecoders.scala
@@ -1,32 +1,27 @@
-package com.binance.api.client.domain
+package com.binance.api.client.json
 
+import com.binance.api.client.domain._
 import io.circe._
 
-object DomainJson {
-  def enumDecoder[E <: java.lang.Enum[E]](values: Seq[E]): Decoder[E] =
-    Decoder.decodeString.flatMap { str => (c: HCursor) =>
-      values
-        .find(_.name() == str)
-        .toRight(DecodingFailure(s"$str not among $values", c.history))
-    }
+trait DomainDecoders {
 
   implicit lazy val OrderSideDecoder: Decoder[OrderSide] =
-    enumDecoder(OrderSide.values())
+    JavaEnumDecoder(OrderSide.values())
 
   implicit lazy val OrderTypeDecoder: Decoder[OrderType] =
-    enumDecoder(OrderType.values())
+    JavaEnumDecoder(OrderType.values())
 
   implicit lazy val TimeInForceDecoder: Decoder[TimeInForce] =
-    enumDecoder(TimeInForce.values())
+    JavaEnumDecoder(TimeInForce.values())
 
   implicit lazy val ExecutionTypeDecoder: Decoder[ExecutionType] =
-    enumDecoder(ExecutionType.values())
+    JavaEnumDecoder(ExecutionType.values())
 
   implicit lazy val OrderStatusDecoder: Decoder[OrderStatus] =
-    enumDecoder(OrderStatus.values())
+    JavaEnumDecoder(OrderStatus.values())
 
   implicit lazy val OrderRejectReasonDecoder: Decoder[OrderRejectReason] =
-    enumDecoder(OrderRejectReason.values())
+    JavaEnumDecoder(OrderRejectReason.values())
 
   implicit lazy val AssetBalanceDecoder: Decoder[AssetBalance] =
     Decoder.forProduct3("a", "f", "l")(AssetBalance.apply)
@@ -35,5 +30,4 @@ object DomainJson {
     Decoder.decodeTuple3[String, String, Seq[String]].map {
       case (a0, a1, _) => OrderBookEntry(a0, a1)
     }
-
 }
